Changes to home page, auctions page, view_auction
So the first thing that I did was move the view auction template
to its own file. It is now possible to click any auction and be
taken to the right auction. I went ahead and cleaned up this partial
as much as possible.

Then I added the ajax calls for everything on that page. You can now
bid, and it will update your total on the page. You can flag an auction
and it will flag the auction in the DB. I also installed an ajax
that checks the price every ten seconds to make sure that you're
looking at the most current price.

On the auctions page I made a lot of changes to the layout to make
it look and behave better. Tomorrow, if there's time, I will update
the my_auctions page to match these changes.